basics for breadcrumb are in place
