--- components/atoms/BreadCrumbReact.js
@@ -1,58 +1,0 @@
-import { useEffect, useState } from "react";
-import { useRouter } from "next/router";
-import Link from "next/link";
-
-const convertBreadcrumb = (string) => {
-  return string
-    .replace(/-/g, " ")
-    .replace(/oe/g, "ö")
-    .replace(/ae/g, "ä")
-    .replace(/ue/g, "ü")
-    .toLowerCase();
-};
-
-export default function BreadcrumbsReact(props) {
-  const router = useRouter();
-  const [breadcrumbs, setBreadcrumbs] = useState(null);
-
-  useEffect(() => {
-    if (router) {
-      const linkPath = router.asPath.split("/");
-      linkPath.shift();
-
-      const pathArray = linkPath.map((path, i) => {
-        return {
-          breadcrumb: path,
-          href: "/" + linkPath.slice(0, i + 1).join("/"),
-        };
-      });
-
-      setBreadcrumbs(pathArray);
-    }
-  }, [router]);
-
-  if (!breadcrumbs) {
-    return null;
-  }
-
-  return (
-    <nav aria-label="breadcrumbs">
-      <div>
-        <div className="inline">
-          <Link href="/">
-            <a className="inline">life journeys </a>
-          </Link>
-        </div>
-        {breadcrumbs.map((breadcrumb, i) => {
-          return (
-            <div key={breadcrumb.href} className="inline">
-              <Link href={breadcrumb.href}>
-                <a>{convertBreadcrumb(breadcrumb.breadcrumb)}</a>
-              </Link>
-            </div>
-          );
-        })}
-      </div>
-    </nav>
-  );
-}
